Upgrade actfw_raspberrypi.vc4.display API
